project: clear remaining compiler warnings across the UI and Aep code
- Add explicit static_cast<> narrowing casts (C++ casts in the .mm files) where
  NSInteger/NSUInteger/RSND ids feed 32-bit slots (Friend/Store view
  controllers, MenuMainTask, RandomLoginBonusView), fixing -Wshorten-64-to-32.
- Drop redundant stringWithString:/initWithString: literal wrappers
  (StoreMainViewController, StoreAcvManageViewController, RewardNetworkWebViewController).
- Declare RandomLoginBonusView's -init as the NS_DESIGNATED_INITIALIZER so its
  -initWithFrame:/-initWithCoder: funnels stop tripping the designated-initializer
  diagnostics.
- Compare the DownloadMain visitor delegate through id to silence the
  distinct-pointer-types warning (CheckerMusicViewController).
- Mark PurchaseManagerDelegate's -finishRequest: optional: only the request
  delegate (AppDelegate) implements it, so PurchaseStore now conforms.
- Pass the quad command's nColorMul colour argument via static_cast<uint32_t>
  (AepOrderingTable).

+++ Project/FriendReplyViewController.mm
@@ -301,7 +301,7 @@
             self.tableView.tableHeaderView = _headView;
         }
         self.tableView.scrollEnabled = (count != 0);
-        [[DownloadMain getInstance] setFriendRequestedCnt:count];
+        [[DownloadMain getInstance] setFriendRequestedCnt:static_cast<int>(count)];
     } else {
         errorMessage = @"エラーが発生しました。"; // best-effort: cf_Ok01YWeW0_0W0_00
     }
@@ -340,7 +340,7 @@
                     self.tableView.tableHeaderView = _headView;
                 }
                 self.tableView.scrollEnabled = (count != 0);
-                [[DownloadMain getInstance] setFriendRequestedCnt:count];
+                [[DownloadMain getInstance] setFriendRequestedCnt:static_cast<int>(count)];
                 break;
             }
         }
